requirements.txt
altgraph==0.17.4
Brotli==1.1.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cramjam==2.8.3
darkdetect==0.8.0
etcpak==0.9.9
fastparquet==2024.5.0
fsspec==2024.9.0
greenlet==3.0.3
idna==3.10
lz4==4.3.3
numpy==2.1.1
packaging==24.1
pandas==2.2.2
pefile==2024.8.26
pillow==10.4.0
pyfmodex==0.7.2
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
pyqt-resource-helper==0.0.14
pyqt-toast-notification==1.3.2
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
PySide6==6.7.2
PySide6_Addons==6.7.2
PySide6_Essentials==6.7.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32-ctypes==0.2.3
QtPy==2.4.1
requests==2.32.3
shiboken6==6.7.2
six==1.16.0
SQLAlchemy==2.0.34
tabulate==0.9.0
texture2ddecoder==1.0.4
typing_extensions==4.12.2
tzdata==2024.1
UnityPy==1.10.18
urllib3==2.2.3
